pyproject.toml
[tool.poetry]
name = "surveillance"
version = "0.1.0"
description = "Motion tracking and notification using opencv and telegram"
authors = ["dunefox"]

[tool.poetry.dependencies]
python = "^3.7"
"python-telegram.bot" = "^12.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
